requirements.txt
Django==4.0.3
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==4.0.1
